Migrate the Grid gizmos to the generic registry system
A grid's rows and columns are not grabbed at a point: any spot along an edge will
do, and the band the edge occupies counts too. So a shape can now supply
`hover_distances` in place of measuring to its handle positions, which is the
same escape hatch the other hooks are -- the generic layer keeps arbitrating
between overlapping gizmos, it just no longer assumes the thing being aimed at is
a point.

Dragging the top edge upward also has to move the layer. The node builds its grid
downward from the origin, so without a matching translation the new rows would
appear at the bottom and the edge would slide out from under the cursor. `drag`
therefore returns a `DragWrites` rather than a bare list of inputs, carrying an
optional transform alongside them.

Dragging past the last row does not stop at one either -- the grid turns inside
out and the opposite edge takes over -- so the gesture re-anchors its origin the
same way the arc re-anchors its baseline.

Rows and columns are declared as dials, but a declaration that brings its own
drag is hosted by the general handle gizmo rather than the dial: the dial is the
narrow one, a count stepped by horizontal drag, and none of that applies here.

The grid's edge geometry -- four edges across two layouts, rectangular and
isometric -- moves to `grid_shape.rs`. It is the grid's own geometry rather than
gizmo machinery, and it is the same shape of split as the star's snap radii.

The spacing gizmo the registry used to declare is dropped. It was a position
gizmo that was never built, and spacing is a two-axis value with no obvious
handle on the shape.

With this the `shape_gizmos` module is empty and deleted. All seven shapes now
run on the registry.
