+++ editor/src/messages/tool/common_functionality/gizmos/mod.rs
@@ -2,4 +2,3 @@ pub mod generic_gizmos;
 pub mod gizmo_behaviors;
 pub mod gizmo_manager;
 pub mod gizmo_registry;
-pub mod shape_gizmos;
