--- editor/src/messages/tool/common_functionality/gizmos/shape_gizmos/mod.rs
@@ -1,1 +1,0 @@
-pub mod grid_rows_columns_gizmo;
